feat: implement handler for user task CREATING intent
to completely overwrite old user task documents in the case of migration.

- refactored acceptance tests to be parameterized
- moved acceptance tests for user task PIM to own clasee
- added test case for embedded form to external reference process instance migration

relates to https://github.com/camunda/camunda/issues/44911
